Update patch set 1
Patch Set 1:

This is apparently a common error with PowerVR GPUs. I remember reading an explanation of why it ("it" being "the soft implementation of some GL extensions takes priority over the target's blob") happened with the SGX530 (probably the same justification as the 540), and it made sense. I wish I remembered what it is :)

In any case, I never did this with my PVR devices, because I really don't like the idea of removing the fallback implementation. I'm not even sure the blob supports all the extensions (but that's easy to find, just compare with/without logcats to see if there's any missing); and I never saw a vendor (including, now, Google) doing it in their own PVRs...

Patch-set: 1