Update patch set 1
Patch Set 1:

Ricardo I also don't have any idea of why is this implemented, but as you said is selected sofware before hardware GPU on this device that is an error. From my experience on nuxus s that makes some games, system apps to be slow and buggy.

On nexus s I don't see missing or erros in log after removing this, but what I see is a better performance. That the reason of this change

Patch-set: 1