Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The phone (Maguro) has become 'stuttery' since using this kernel.  I am not overclocking the CPU as I know not all devices can handle overclocking and I am still noticing that the device is not as smooth as it was with the previous kernel.

Patch-set: 1