Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Also, what do we get by upping to 3.0.14?
I think we shouldn't get too out of control for the CM kernel.
We should base the CM kernel on the android-omap-tuna-3.0
branch and let google merge in from the mainline kernel.
We can cherry pick fixes/features and modify the defconfig slightly, but that should be it.
We have a good possibility of breaking things by not doing that.

Patch-set: 1