Update patch set 1
Patch Set 1:

I agree, interactive should be the default.
I think what we need to do is branch the kernel tree into the Cyanogen github so we can track actual changes to the kernel, rather than guessing whats in the binary uploads.
I'm happy to manage that repo if you need a volunteer.

Patch-set: 1