Update patch set 1
Patch Set 1:

Now is upgraded to 3.0.16 and is a long term upgrade that is not going to make any bad at all.

There you have cifs, ntfs... and I don't see the reason to evolute that kernel. If we just wait for what does google with his device, the non sense to build CM. We folow all aosp work and we don't evolute in any branch. In my opinion we should make aour best in all the parts, not only folowing what google does.

GPU overclock is safe, CPU overclock is up to the user.

About the complain on governor, the default google governor is ondemand. I know is compiled as interactive, but look at the GN boot sequence, and you will se is rewrited at boot to change to ondemand, so you using interactive, are setting the kernel in a diferent governor than it is setted by default.

Patch-set: 1