Update patch set 1
Patch Set 1:

David and jonathan, feel free to fork my github code and pull all your suggest to contribute into the kernel. This is not at all an only person work, and all the help will be wellcomed.

The reason I now use gerrit for kernel is not generating a lot of non sense stuff for other users, and also the possibility to create branches to make all the patches in a clear way.

Patch-set: 1