[Tuna] Add product specific definition to HDCP symlink
Add product specific definition to the HDCP symlink creation.

Without this protection performing a make after lunch for other products
will force the inheritance of this symlink into the build.
When creating the ota and update packages for other targets these
symlinks are not created.

This will cause other products to fail HDCP symlink association since
the symlink was not created for the target product.

Change-Id: I632be3822f0a6e281a9083e9474559eda4269a50