Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Lets take crespo as example, since is also a google nexus device. We have more than on booloader accepted... But this in fact cames from google, this file is like google once. For some reason google is suggesting to use this especific bootloader and radio on 4.0.4 So I'm afraid about accepting old bootloaders and radios and this can produce bad behaviour of cyanogenmod on 4.0.4
So are you sure this older bootloader is not going to produce a bad behaviour now or in the future? So the question is why did they removed it from the configuration on tuna?
I would like to be sure of that change before submiting

Kalim

Patch-set: 3