Update patch set 1
Patch Set 1: Do not submit

changing the buildchain from the original from google needs a real data of better results, like benchmarks or other real testable data.

Remove original debugging is only needed when the ROM and kernel is fully finished, not at that moment that we are still testing and so.

I preffer to maintain the config as google did. tuna_defconf that is an incremental file of only the changes.

Really you need zcache with maguro?

All the changes should be pushed (pull request) into https://github.com/CyanogenMod-SAMSUNG-Kernels/android_kernel_samsung_tuna and if accepted there will be compiled and sended to gerrit. We are working into move all the kernel into gerrit.

Kalim

Patch-set: 1