Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I agree with Kalim and his reasoning.

Changing toolchain and optimizations, with the corresponding microchanges to the generated code, potentially makes for hard to diagnose and hard to track down bugs and variations. Such a step, IMO, needs thorough and prolongued testing.

It would thus be good to have an accessible venue for such testing to happen. A not-to-be-merged-soon gerrit change like this here is fine, I think. Also maybe adding the varied config into the "base" source repository, e.g. as a tuned_tuna_defconfig, so people can reproduce and make further experiments. Finally, an easy to install update.zip downloadable on the web for end users to test, along with a SEPERATE XDA thread to gather results.

There is one change here which might be generally applicable, nr. 3 - leaving out ext2.

Regarding zcache, I'm also somewhat doubtful. zcache trades off memory / restarting not-recently-used apps, against additional CPU usage for compression and general bookkeeping. Given the amount of RAM we have on galaxy nexus, I don't think that tradeoff will be a good one. It is also hard to quantify, because you'd need a well loaded system (not normal in my daily use at all) and then a good benchmark suite tugging along.

Patch-set: 1