Update patch set 1
Patch Set 1:

There's an option for zram in performance options.  I think this may have meant to be called zcache because it was ported from CM7s compcache.  Anyway, my point is, if there's an option for it in settings should it not be implemented in the kernel?  It can be made to default to off.  If not the settings option should probably be removed.

Patch-set: 1