Update patch set 1
Patch Set 1:

Sorry, don't think I explained myself very well.  I know the difference between zram and zcache.  What I was saying is that  when the performance settings were merged to ics from CMparts:

https://github.com/CyanogenMod/android_packages_apps_Settings/commit/6dca7c4752f023f39a60538f45a1e99e23225468

https://github.com/CyanogenMod/android_frameworks_base/commit/5ae808cc55144e5218f74d9bf11846a81ec476ea

Compcache was renamed ZRAM. I think this is wrong and should be ZCACHE.

If there's an option in performance settings for ZCACHE, it should be enabled in the kernel.  If the consensus it that ZCACHE is not required and shouldn't be in the kernel then the option should be removed from performance settings because it wouldn't be doing anything.

Patch-set: 1