Update patch set 1
Patch Set 1:

Steve and other members are working hard in having Tuna kernel in gerrit. I will be very pleased and for sure the rest of the team to work on those patches one by one and discussing all toguether on the benefits of them soom. For sure most of you have some great ideas we can merge to improve this work and have a better release day by day. For sure with the collaboration of Menno, jonathan, Daniel, Patrick... we will improve more, better and faster soon...

About zcache/zram will be no improvement for this device. We have enought RAM and we don't need that patch that will be only CPU waste on GN devices.

KalimAz

Patch-set: 1