Update patch set 1
Patch Set 1:

Agreed, and I look forward to that...
In regards to zcache. Again, please dont confuse it with Zram.
This is more about I/O performance.
Here are my stats from my GN after just being on for 12 hours.

flushes:        33564
puts:           36780
succ_gets:      4429

So there is an I/O performance boost here.
I will do some perf. tests with and without Zcache and report.
I'm not just trying to add a feature for sake of adding a feature, I want to get something out of it.

Patch-set: 1