Update patch set 1
Patch Set 1:

Jonathan,

Please don't feel like we are ignoring your work, I'm not, otherwise I will just ignore all that conversation and topic.

As we are all working on the same topic ( Tuna Kernel ) we need to talk about the benefits of all that patches. For example, I see a lot of non benefits of using a new toolchain since I did that long time ago. Imaging our buildboots, using googles toolchain we are compiling all the devices at once and is all ok, so is difficult to manage to have a specific toolchain for each compilation/device/system/kernel... But there is a point that can convince me and the others. Just try to make a compilation with this toolchain and google one with the same config and flags on both, do you see any benefit? I maen, I have done a test with some bench apps and no appreciable diference and remember that all the ROM and the privative blobs from google are compiled with the same toolchain. Can you try this?

About the ext2, I aggree we can remove it.

About the zcache. Can you tell me what is the benefit? I'm checking it here and trying to undertand when this is helpfull on our devices. We are not acceding to a harddrive so the beneffit of reading this pages compressed from RAM and decompress or reading from the system has to be mesured. I have not already done any test of this, we need to prepare a test to see the result of this. http://lwn.net/Articles/397574/

I will analyze all the improvements suggested by anyone here, will never ignore them.

KalimAz

Patch-set: 1