Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Daniel you need to enable also the camera adding the image in one non visible part of the screen. Look at this https://github.com/KalimochoAz/IcsTorch
I have the same issue in crespo. The Question is, we work on creating the surface ( if you creste the surface for the camera image, torch will work ) for the camera on this crazzy app or we think in making a completely new app?

Patch-set: 1