Update patch set 1
Patch Set 1:

I agree with the revert. generally, if there is danger of damage to user devices.

Just one question: did you discuss / wouldn't it be good, to just revert this change, i.e. not provide the Settings part, but keep the kernel side change for the benefit of those who  are ready to take the risk and fondle the /sys/modules/... setting themselves?

Patch-set: 1