Update patch set 1
Patch Set 1:

I haven't tried this change yet but will it correctly handle mounting NTFS partitions, ie. mount it with a umask that allows other than root to read it?

The old behaviour was it popping up a notification that couldn't be dismissed after disconnecting or unmounting in terminal so should be better regardless.

Patch-set: 1