Update patch set 1
Patch Set 1:

The NTFS Linux kernel driver still only supports reading, right? That's the one used currently, not NTFS-3G. That's fine in most cases though as I think the primary use case is watching HD movies from it.

Yeah, I only mentioned the notification because you couldn't clear it before and it didn't properly react to changes. The notification in itself is a good thing.

Unfortunately I can't test this change since my USB port on my GN is glitchy. Basically it only works with my Nexus One charger (not the GN charger though) and a cheap converter I bought from DX. Apparently it's a common problem for some Samsung devices but I have yet to hand it in for repair. :(

Patch-set: 1