Update patch set 1
Patch Set 1: Do not submit

I don't care if it was discussed or not, we made a rule of this: Nobody merges code that isn't either his/hers, from another CM repository, or from upstream.

If it comes from downstream or from an application, it's either submitted by the author, or explicitly signed off by the author as a reviewer here. No exceptions.

Patch-set: 1