Update patch set 1
Patch Set 1:

I don't understand what is the issue. This code is not from codeworkx and the kernel modif is already send to codeworkx autorship.
This has being already talked with him yesterday morning:


10:04:05 <+KalimAz> codeworkx: What if I make an slider to control the powerfrom 0 to 100% in parts? insteed of 0 to 127?
10:04:16 <@codeworkx> that's what i wanted to do          
10:04:21 <@codeworkx> but budy with touchlights
10:04:24 <@codeworkx> busy*
10:04:38 <+KalimAz> I can do relatively quick
10:04:45 <+KalimAz> np
10:04:46 <@codeworkx> and i can kang it quickly

But codeworkx is added to this commit to have his feedback before merge.

So what I'm missing there?

Patch-set: 1