Update patch set 1
Patch Set 1:

The issue is that parallel implementations will not avoid multiple very similar but not identical interfaces for the same function.
If we say this was genuinely created from scratch independently, and if this is acceptable to you, then how does gerrit change anything on the original problem if all it takes is for someone to not use the code verbatim but reimplement it changing sysfs locations for no technical reason, each and every time?

Patch-set: 1