Update patch set 2
Patch Set 2:

The problem we were trying to address is the case where someone makes a fairly complex feature and another developer cherry-picks it into their own repository and makes changes which cause incompatibility. Parallel implementations which are totally different, well that's harder. I don't think CM can really help here. As someone who's been involved in open-source work for a long time, I know that they either happen because one developer was not aware of another's work, or because of social/personality disagreements with one another.

Should you standardize on an API? Maybe. But it's not like any of the guys arguing with this are throwing code at us. Did anyone steal anyone's work? It certainly doesn't appear that way.

Patch-set: 2