Update patch set 2
Patch Set 2:

By the way Jef, what i was thinking of with the opening of changes on gerrit is to help the awareness.
Also, let's face it, that is not a whole lot of code.
If CM has no indicator the original author has any intentions to ever "allow" merging, you would lose out on a whole lot of easy additions.

Use the devrel page, have conditions for inclusion and compatiblity.
Make it known that if a custom kernel dev wants to ensure all the rom bits will revolve around the "original implementation", they must open a change on gerrit early and that they will have a limited lifespan, so they can't be abused as useless/malicious blockers.

Then you have a chance to simply link to other kernels of other devices if it's been done before and "cool you want to do this for device XYZ, but it must match the existing interfaces already in use on CM devices DEF and GAB.

This should make things pretty smooth for everyone if the rules are known and followed.

And if you do choose that route you can pretty much say, follow these simple guidelines or GTFO.

Think that's fair and things will go back to smooth.

Patch-set: 2