Update patch set 2
Patch Set 2:

It's a pretty sure thing that if someone has a cool feature in their kernel for CM, it will be the same for any other device that wants to use it.

In the past there were incidents where someone cherry-picked a change and made modifications to it without involving the original authors. But I think this is just a case where two people made a similar feature independently (or maybe one person had the idea first, but didn't want to submit it so someone else wrote their own).

And I'm +++++ in favor of using Gerrit as a place to flesh out ideas and get people on the same page- that's exactly what it's for and there are a lot of eyes here. We really just want to see good stuff get to as many people as possible.

Patch-set: 2