Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(1 inline comment)

Had a quick look, and the logic seems incorrect. Based on the old code, I'd move the sysfs write outside of the if to deduplicate it, keeping the on ? : comparison, and use the macro instead of the hardcoded path.

Patch-set: 1