Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Why not just make the code a little smarter and detect this stuff automatically?

Soon we will be using boostpulse in many more places in the framework to improve interactivity in ways that can't be done from the kernel alone.

Patch-set: 1