Update patch set 1
Patch Set 1: No score

I submitted that patch already. I took it from AOKP source code but it looks like now someone replaced it with something that is not working.

And next I noticed the commit of adding the Power HAL library.

I have thoroughly tested and monitored my Nexus using this. I have no errors in kmsg or logcat.

So why would we add bloated code if it can be done in a simple and clean way? The Nexus S doesn't have this Power HAL library as well as well as a few other devices running Jelly Bean I looked up.

Patch-set: 1