Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Is it possible to make the GPU OC array with the GPU frequencies universal? By this I mean reading the available GPU OC frequencies from the sysfs interface??? Please add this if it is possible.

Looks ok to me.

Patch-set: 2