Update patch set 1
Patch Set 1:

Does the interface between the frameworks and the power HAL have any methods for configuration?  The power HAL isn't too well documented and I have yet to see an example of a configurable one or even a way to configure it.  I've avoided power HALs on I9100/I777/N7000 (and codeworkx has -1ed them on I9300) due to this inflexibility.

Patch-set: 1