Update patch set 1
Patch Set 1:

I'm not going to -2 this because ultimately it's Rafael's device, but removing the Power HAL because the overclockers don't know what to do is not a good idea.

The current Power HAL is very stupid, userspace sends hints when there are touch events, and when the screen turns on and off. The API supports arbitrary data in these calls but currently this isn't used.

I am currently working on a project that extends the Power HAL to do more things such as manage CPU hotplug to improve interactivity. I'll also be adding support to manage the sysfs permissions correctly when things change via uevents.

Ultimately, we should be able to get rid of device-specific Power HALs and centralize on one implementation since they are mostly the same right now.

Patch-set: 1