Update patch set 3
Patch Set 3:

Before this is submitted, I think it would be nice if users were made aware of the differences between what "fast" and "max" are through some sort of included documentation snippet in the UI. Short of reading the source or this Gerrit, it would be rather hard for anyone to actually discern the differences between the two (other than the "feel" factor and that's always bad news).

Semi off topic, but I've been following the CM Gerrit for a while and commits (as well as working with the source on my own) and I noticed that no one really documents their code as far as at least leaving doc blocks with a comment as to what a function/method does. Mostly the comments seem to be relegated to what Google had previously added themselves. 

A minimal level of documentation (at least a doc block with a description) would help those I think that are still getting acquainted with the code (or don't work with it on a daily basis) to avoid having to unnecessarily read through parts that are unrelated to what one is trying to patch/change.

Patch-set: 3