Update patch set 3
Patch Set 3:

The stock ROM with maximum brightness values ??equal to half the setting in CM (200). Exceeding these values ??represent a higher battery consumption, and the possibility of degrading the display with intensive brightness. It is therefore advisable to warn the user of this characteristic.

Of course, the decision to include or not this warning is yours.

Patch-set: 3