Update patch set 1
Patch Set 1: No score

Hmm. I see. The only thing I can think of, even though it works now, is for the future, incase any changes are made to the one included in the DSP Manager source. That would lead to having to add the changes to this one again. Also, with the the way the audio_effects.conf the way it is now, I'm pretty sure it's identical to the one included with the DSP Manager source, although I'd have to double check. Mainly submitted this because, what's the sense in having the same file in two different places, right?

Patch-set: 1