Update patch set 3
Patch Set 3:

Since this isn't actually used for any of the devices, I don't see a need to merge this. Adding it in just makes for another commit to work with when merging in upstream changes. Since the real issue is fixed by the device-specific changes, I'd rather see this abandoned.

Patch-set: 3