Change CM color-multiplier preference/preset for Tuna
While I realize that the CM philosophy is high-color/low-brightness, a
multiplier value of 350 (1750000000) is supposed to be unsafe for AMOLED
displays. It even states "Values higher than 200 are not recommended" in
the multiplier settings menu, yet 350 is the CM setting. I've changed it to a
value of 250 instead, as a sort of compromise. That way, it still has brighter
colors, while not going "radically" above what is recommended. The real problem
lies in poeple being unaware that this value is set. I've seen reports of people
flashing new builds, and the CM values are automatically set for whatever reason.
This has even happened to me a couple times in the past. Some individuals were
actually quite upset about this, concerned that their display was ruined because
of it. It's a very minor, yet worthy change in my opinion.

Change-Id: Ie704c79c8c4a01e868e3cf0ec1ea77a9a847b46f