Update patch set 1
Patch Set 1:

As Dewayne and Sean said, it would be best to just add that to tuna's device.mk to avoid having to change 3 other repos (maguro, toro, and toroplus) as well.

Screenshots are working just fine, however the preview animation (where it "shrinks" the screen when the screenshot is taken) is completely garbled, but that was an issue before this so it isn't relevant anyways.

Patch-set: 1