Update patch set 2
Patch Set 2: Code-Review+1

Looks great, however one thing that would be nice is a change to libs/gui/Android.mk in frameworks/native to automatically set the FORCE_SCREENSHOT_CPU_PATH if e.x. "TARGET_FORCE_SCREENSHOT_CPU_PATH" (somehow condense that of course) is set in the BoardConfig, rather than having that cflag specified globally. That is out of the hands of this particular patch though and can be done later, so +1.

Patch-set: 2