Update patch set 1
Patch Set 1: Code-Review-1

Sorry for the late reply, I am pretty busy lately.
As for this patch, due to the recent ffmpeg changes, we no longer need to copy in the ffmpeg codecs xml, as it's automatically copied in - which is why change 103802 removed the relevant line from device.mk. However, in order to still use the ffmpeg codecs, we still need to reference them from media_codecs.xml - which is what your patch set removes.
Is there something I'm missing here? I admit I'm not fully up-to-date with CM's recent changes, but I think this is how it works currently :)

Patch-set: 1