Update patch set 1
Patch Set 1:

Without the low_ram flag, the Settings app used ~80 MB of RAM - with it, it only uses ~50 MB. This is true for all apps - while I didn't get around to actually check what it tweaks (some dalvik values?), with the combination of this commit, it helps a lot! Maybe you could test if the same happens on the Moto G too? Ever since using this combination, my memory problems disappeared :)

Patch-set: 1