Update patch set 3
Patch Set 3:

I do have both toro and toroplus running on cm12.1, you may have seen my unofficial builds over on xda.  I've been using this SELinux patch for quite some time on them.

However, it works equally well to simply remove the Verizon and Sprint proprietary apk's.  This is how we are handling it on Marshmallow, as the proprietary apk's are just crashing there.  Credit to @MWisBest for figuring that out.

Patches needed to get toro(plus) working? that depends:  We either need a few to frameworks/opt/telephony and frameworks/native, or we need the modified libsec-ril_lte.so files from MWisBest.  I'm not sure how the proprietary vendor repo is handled for cm, is using the modified libs an option?

Patch-set: 3